+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/ContainerExecDecorator.java
@@ -42,7 +42,6 @@ import org.apache.commons.io.output.TeeOutputStream;
 import org.csanchez.jenkins.plugins.kubernetes.pipeline.proc.CachedProc;
 import org.csanchez.jenkins.plugins.kubernetes.pipeline.proc.DeadProc;
 
-
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import hudson.EnvVars;
@@ -77,6 +76,7 @@ public class ContainerExecDecorator extends LauncherDecorator implements Seriali
     private static final String COOKIE_VAR = "JENKINS_SERVER_COOKIE";
 
     private static final Logger LOGGER = Logger.getLogger(ContainerExecDecorator.class.getName());
+    private static final String DEFAULT_SHELL="/bin/sh";
 
     private transient KubernetesClient client;
 
@@ -93,6 +93,7 @@ public class ContainerExecDecorator extends LauncherDecorator implements Seriali
     private EnvVars globalVars;
     private FilePath ws;
     private EnvVars rcEnvVars;
+    private String shell;
 
     public ContainerExecDecorator() {
     }
@@ -105,6 +106,7 @@ public class ContainerExecDecorator extends LauncherDecorator implements Seriali
         this.containerName = containerName;
         this.environmentExpander = environmentExpander;
         this.ws = ws;
+        this.shell = DEFAULT_SHELL;
     }
 
     @Deprecated
@@ -196,6 +198,14 @@ public class ContainerExecDecorator extends LauncherDecorator implements Seriali
         this.ws = ws;
     }
 
+    public String getShell() {
+        return shell == null? DEFAULT_SHELL:shell;
+    }
+
+    public void setShell(String shell) {
+        this.shell = shell;
+    }
+
     @Override
     public Launcher decorate(final Launcher launcher, final Node node) {
         return new Launcher.DecoratedLauncher(launcher) {
@@ -204,8 +214,6 @@ public class ContainerExecDecorator extends LauncherDecorator implements Seriali
                 LOGGER.log(Level.FINEST, "Launch proc with environment: {0}", Arrays.toString(starter.envs()));
                 boolean quiet = starter.quiet();
                 FilePath pwd = starter.pwd();
-
-
                 List<String> procStarter = Arrays.asList(starter.envs());
                 List<String> cmdEnvs = new ArrayList<String>();
                 // One issue that cropped up was that when executing sh commands, we would get the jnlp agent's injected
@@ -215,7 +223,7 @@ public class ContainerExecDecorator extends LauncherDecorator implements Seriali
                 // Currently, build level properties will be provided by the Run Context anyways.
                 boolean javaHome_detected = false;
                 for (String env : procStarter) {
-                    if (env.contains("JAVA_HOME")) {
+                    if (env.equalsIgnoreCase("JAVA_HOME")) {
                         LOGGER.log(Level.FINEST, "Detected JAVA_HOME in {0}", env);
                         javaHome_detected = true;
                         break;
@@ -309,7 +317,7 @@ public class ContainerExecDecorator extends LauncherDecorator implements Seriali
 
                 ExecWatch watch;
                 try {
-                    watch = execable.exec("/bin/sh");
+                    watch = execable.exec(getShell());
                 } catch (KubernetesClientException e) {
                     if (e.getCause() instanceof InterruptedException) {
                         throw new IOException("JENKINS-40825: interrupted while starting websocket connection", e);